localpasswdmngr/requirements.txt
cryptography-44.0.2 
mariadb-1.1.12

